requirements.txt
httpx>=0.27.0
psycopg2-binary>=2.9.9
numpy>=1.26.0
python-dateutil>=2.9.0
pydantic>=2.7.0
feedparser>=6.0.11
PyGithub
